Refactor items from array of strings to array of objects with unique IDs
Refactor items from array of strings to array of objects with unique IDs

Updated the items array in App.vue to use objects with unique IDs instead of simple strings. This change improves the data structure by allowing for more complex item attributes and better key management in the v-for loop.