Add newItem ref and input field for adding items to the list
Add newItem ref and input field for adding items to the list

Introduced a new reactive reference `newItem` to manage user input. Added an input field in the template bound to `newItem` with v-model for adding new items to the list. This setup prepares for future implementation of item addition functionality.